fix(docs): stabilize Fumadocs type cache keys
