+++ .github/actions/nextjs-cache/action.yml
@@ -49,13 +49,16 @@ runs:
         BENCHMARK_CACHE_KEY: ${{ inputs.benchmark-cache-key }}
         CACHE_GENERATION: ${{ inputs.cache-generation }}
         CACHE_SCOPE: ${{ inputs.cache-scope }}
+        DEFAULT_BRANCH: ${{ inputs.default-branch }}
         LEGACY_BUILD_CACHE_KEY_SUFFIX: ${{ inputs.legacy-build-cache-key-suffix }}
         NEXTJS_COMPATIBILITY_HASH: ${{ hashFiles('bun.lock', 'package.json', 'bunfig.toml', 'docs/package.json', 'docs/next.config.mjs', 'docs/postcss.config.js', 'docs/tsconfig.json', '.github/actions/setup/action.yml') }}
-        FUMADOCS_COMPATIBILITY_HASH: ${{ hashFiles('bun.lock', 'docs/package.json', 'docs/tsconfig.json', 'docs/components/type-table/generator.ts', 'docs/source.config.ts', 'docs/registry/react/**', 'docs/registry/lynx/**', 'packages/css/**', 'packages/react/src/**', 'packages/react-headless/**', 'packages/lynx-react/src/**', 'packages/stackflow/src/**') }}
       run: |
         nextjs_namespace="${RUNNER_OS}-nextjs-v2-"
         nextjs_legacy_namespace="${RUNNER_OS}-nextjs-"
-        fumadocs_namespace="${RUNNER_OS}-fumadocs-typescript-v1-"
+        fumadocs_namespace="${RUNNER_OS}-fumadocs-typescript-v2-"
+        # 외부 Actions cache와 내부 JSON generation이 같은 호환성 기준을 사용한다.
+        # 이 해시는 setup 단계에서 복원되는 lib와 .ultra.cache.json 같은 생성물을 제외한다.
+        fumadocs_compatibility_hash="$(bun -e 'import { typeTableCacheCompatibilityHash } from "./docs/components/type-table/cache-compatibility.ts"; process.stdout.write(typeTableCacheCompatibilityHash)')"
 
         if [ -n "$BENCHMARK_CACHE_KEY" ]; then
           nextjs_namespace+="benchmark-${BENCHMARK_CACHE_KEY}-"
@@ -66,13 +69,13 @@ runs:
         fi
 
         nextjs_namespace+="${NEXTJS_COMPATIBILITY_HASH}-"
-        fumadocs_namespace+="${FUMADOCS_COMPATIBILITY_HASH}-"
+        fumadocs_namespace+="${fumadocs_compatibility_hash}-"
 
         echo "nextjs-primary-key=${nextjs_namespace}scope-${CACHE_SCOPE}-${CACHE_GENERATION}" >> "$GITHUB_OUTPUT"
-        echo "nextjs-restore-prefix=${nextjs_namespace}" >> "$GITHUB_OUTPUT"
+        echo "nextjs-restore-prefix=${nextjs_namespace}scope-${DEFAULT_BRANCH}-" >> "$GITHUB_OUTPUT"
         echo "nextjs-legacy-restore-prefix=${nextjs_legacy_namespace}" >> "$GITHUB_OUTPUT"
         echo "fumadocs-primary-key=${fumadocs_namespace}scope-${CACHE_SCOPE}-${CACHE_GENERATION}" >> "$GITHUB_OUTPUT"
-        echo "fumadocs-restore-prefix=${fumadocs_namespace}" >> "$GITHUB_OUTPUT"
+        echo "fumadocs-restore-prefix=${fumadocs_namespace}scope-${DEFAULT_BRANCH}-" >> "$GITHUB_OUTPUT"
 
     - name: Restore Next.js cache
       id: restore-nextjs
